move headerops.get_stream_handlers to fileio
